putting files under python.
